feat: cd to workspace folder before executing yapf
This is necessary to have yapf honor `.yapfignore` or `pyproject.toml`
files, which it reads from the directory that it is executed in. With
the change, the call to `yapf` also passes `.` as the target for
formatting; otherwise, relative paths specified in the ignore patterns
do not match against the absolute target path.

This resolves #7.

Signed-off-by: Ingo Müller <ingomueller@google.com>